refactor front /info-pack
